add example to each Control docs section
