+++ packages/core/src/components/forms/switch.md
@@ -3,6 +3,8 @@
 A switch is simply an alternate appearance for a [checkbox](#core/components/forms/checkbox) that
 simulates on/off instead of checked/unchecked.
 
+@reactExample SwitchExample
+
 @## CSS API
 
 @css pt-switch
